Updated the Quest widget with two new features:
The first is that the Miscellaneous objectives can now be activated and deactivated by double clicking on them.

The second feature is a Show on Global Map button. It will work by either showing the quest objectives (for normal quests), or by selecting the miscellaneous quest objective.

I had to update the QuestMarker object in the GlobalMapWidget to store form id information so that the iwcCenterOnQuest method I created could convert form ids to quest marker ids.