Initial commit of the Perks widget. This should act like the official app with two notable exceptions:
1) The widget shows number of ranks on the left of the Perk name instead of in the perk title,

2) It does not show the perk animation.

I went a little fancy and created a Star graphic. I have placed it in the master Res folder so other widgets can use it.

And the last feature aside from showing the perks and their descriptions, are buttons that let you cycle between the levels of a perk. I have opted that upon a perk rank update, if you have the perk selected, it will not update the stars or description right away. This is so you have an uninterrupted view of the description.