Removed some comments from PerksWidget and QuestsWidget (no longer required).
Initial commit of the Player Statistics widget. This includes information on the players limbs, their total damage output, total damage resist, and special stats. The Damage/Resist section is using funky SVG graphics that change color based on the in game pipboy color. I also made sure this widget would load correctly when the player is in the main menu.

The PerksWidget is also now using fancy SVG graphics that change color based on the in game pipboy color.