A few changes this time around.
First and foremost, since I was using the same code for managing and showing icons, I moved that code off to its own class file. I also moved all of my icons to one place (widgets/shared/res/) and the PipboyIcon class loads from these files specifically.

As a result of the above change, I updated the StatsWidget and the PerksWidget.

Lastly, I created a Workshop widget to show the user which workshops they have open. It will show the list and bold any workshop that is having problems. Upon clicking on a workshop, it will load its details on the right side of the widget. Aside from having icons for each resource, I have also included warning symbols beside each. They will hide (or change to the same color as the background) when a resource does not have a negative rating.