Initial commit of the Game Statistics widget. This simple two table widget shows the game stat sections on the left, and upon clicking a section title, it will show the stats on the right.
Because of size weirdness, it shows the value first, then the text describing the statistic. That way I can resize the numbers column to fit the contents, and let the rest of the table be for the text. And since there is a slider, you can show or hide the text.