Cleaned up my code so everything is following the same guidelines and format.
Converted all UI update calls to be Qt Signals.

Found a bug with the Stats widget where if I connect during the main menu and load a game, the Data's were not being updated. So I went back to a single 4 levels deep call to PlayerInfo. Everything updates now, but I'm afraid we might be seeing problems again with Akamal and the HotKey healing lag. Here's hoping no problems...