Initial commit of the Auto Doc widget.
This is, by most definitions, a cheat widget. It literally handles the health and well being of your character.

I have done as much as I could to ensure that this mod is not super aggressive on reads to the Data, and processing for things like the main logic and UI updates. However with that in mind, if you want responsiveness, it needs to do the calls that it does.

The Auto Doc handles:
* Stimpak injections (yeah, on foot it WILL override your actions and run the animation)
* Med-X application when hurt, when you take a Stimpak, and at 3/4, 1/2, and 1/4 health
* RadAway injections when radiation has been detected (if you started with rad damage, it won't see it), or when taking a Stimpak and it heals nothing
* Rad-X application when radiation has been detected, or when taking a RadAway

For this commit, it doesn't remember your settings. Next commit will have that feature.