Updated my widgets so that they do not process or update the UI when they are not visible. This should help save some processing power on weaker machines.
I also learned while playing that when I changed the Auto Doc to use a generic class for every drug, I also disabled the preventative measure stopping the stimpak injection loop when irradiated too much. This has been re-enabled.

And lastly, there was a situation that if radaway use was disabled, and you were irradiated, the Rad State would keep updating and show that you were no longer irradiated (when you clearly just were). I updated it so it understands that radiation goes away when you take radaway. It is unknown at this time if visiting the doctor will also have the same effect. It is more precise to say that the irradiated state goes away when it detects the negative rads effect of type 41 (which also re-enables the use of stimpaks due to the above mentioned freeze).